pyproject.toml
[tool.poetry]
name = "osint_tools"
version = "0.1.4"
description = ""
authors = ["Alexander Slessor <alexjslessor@gmail.com.com>"]
readme = "README.md"
packages = [
    {include = "osint_tools"}
]

[tool.poetry.dependencies]
python = "^3.8"
requests = ">=2.11"
pydantic = "^1.10.1"
motor = "^3.0.0"
feedparser = "^6.0.8"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
httpx = "^0.23.0"
asgi-lifespan = "^1.0.1"
pytest-asyncio = "^0.19.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
